pipeline/requirements.txt
absl-py==1.3.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
biopython==1.79
black==22.1.0
bleach==5.0.0
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
colorcet==3.0.0
contourpy==1.0.7
cutadapt==3.7
cycler==0.11.0
Cython==0.29.33
debugpy==1.5.1
decorator==5.1.1
dnaio==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
fastjsonschema==2.15.3
flatbuffers==22.12.6
fonttools==4.29.1
gast==0.4.0
getch==1.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.6.0
holoviews==1.14.9
hvplot==0.8.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.9.0
ipython==8.0.1
isal==0.11.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.11.0
kiwisolver==1.3.2
libclang==14.0.6
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nbformat==5.4.0
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.2
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
panel==0.13.1
param==1.12.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
plotly==5.8.0
prompt-toolkit==3.0.28
protobuf==3.19.6
ptitprince==0.2.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyct==0.4.8
Pygments==2.11.2
PyHamcrest==2.0.4
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2021.3
pyviz-comms==2.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
stack-data==0.1.4
tenacity==8.0.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xopen==1.4.0
zipp==3.8.0
